dbcleaner for MS SQL Server (only for java)

+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/db/DbCleaner.java
@@ -23,7 +23,7 @@ public class DbCleaner {
     }
 
     public static void clearDatabase_H2(Connection connection, List<String> tablesToSkip) {
-        clearDatabase_H2(connection, "PUBLIC", tablesToSkip);
+        clearDatabase_H2(connection, getDefaultSchema(DatabaseType.H2), tablesToSkip);
     }
 
     public static void clearDatabase_H2(Connection connection, String schemaName, List<String> tablesToSkip) {
@@ -37,20 +37,21 @@ public class DbCleaner {
          */
 
         try {
-            Statement s = connection.createStatement();
+            Statement statement = connection.createStatement();
 
             /*
                 For H2, we have to delete tables one at a time... but, to avoid issues
                 with FKs, we must temporarily disable the integrity checks
              */
-            disableReferentialIntegrity(s, type);
+            disableReferentialIntegrity(statement, type);
 
-            truncateTables(tablesToSkip, s, schemaName, isSingleCleanCommand(type), doDropTable);
 
-            resetSequences(s, type, schemaName);
+            cleanDataInTables(tablesToSkip, statement, type, schemaName, isSingleCleanCommand(type), doDropTable);
 
-            enableReferentialIntegrity(s, type);
-            s.close();
+            resetSequences(statement, type, schemaName);
+
+            enableReferentialIntegrity(statement, type);
+            statement.close();
         } catch (Exception e) {
             /*
                 this could happen if there is a current transaction with a lock on any table.
@@ -76,6 +77,9 @@ public class DbCleaner {
         }
     }
 
+    public static void clearDatabase(Connection connection, List<String> tablesToSkip, DatabaseType type){
+        clearDatabase(connection, getDefaultSchema(type), tablesToSkip, type);
+    }
 
     public static void clearDatabase(Connection connection, String schemaName, List<String> tablesToSkip, DatabaseType type){
         clearDatabase(getDefaultReties(type), connection, schemaName, tablesToSkip, type, false);
@@ -101,11 +105,10 @@ public class DbCleaner {
      * @param doDropTable specify whether to drop tables which is only for MySQL and MariaDB now.
      * @throws SQLException are exceptions during sql execution
      */
-    private static void truncateTables(List<String> tablesToSkip, Statement statement, String schema, boolean singleCommand, boolean doDropTable) throws SQLException {
-
+    private static void cleanDataInTables(List<String> tablesToSkip, Statement statement, DatabaseType type, String schema, boolean singleCommand, boolean doDropTable) throws SQLException {
         // Find all tables and truncate them
         Set<String> tables = new HashSet<>();
-        ResultSet rs = statement.executeQuery(getAllTableCommand(schema));
+        ResultSet rs = statement.executeQuery(getAllTableCommand(type, schema));
         while (rs.next()) {
             tables.add(rs.getString(1));
         }
@@ -114,6 +117,7 @@ public class DbCleaner {
         if (tables.isEmpty()) {
             throw new IllegalStateException("Could not find any table");
         }
+
 
         if (tablesToSkip != null) {
             for (String skip : tablesToSkip) {
@@ -126,6 +130,15 @@ public class DbCleaner {
             }
         }
 
+        Set<String> tablesHaveIdentifies = new HashSet<>();
+        if (type == DatabaseType.MS_SQL_SERVER){
+            ResultSet rst = statement.executeQuery(getAllTableHasIdentify(type, schema));
+            while (rst.next()) {
+                tablesHaveIdentifies.add(rst.getString(1));
+            }
+            rst.close();
+        }
+
         List<String> tablesToClear = tables.stream()
                 .filter(n -> tablesToSkip == null || tablesToSkip.isEmpty() ||
                         !tablesToSkip.stream().anyMatch(skip -> skip.equalsIgnoreCase(n)))
@@ -136,24 +149,56 @@ public class DbCleaner {
                     .sorted()
                     .collect(Collectors.joining(","));
 
+            if (type != DatabaseType.POSTGRES)
+                throw new IllegalArgumentException("do not support for cleaning all data by one single command for " +type);
+
             if (doDropTable)
-                statement.execute("DROP TABLE IF EXISTS " +ts);
-            else
+                dropTables(statement, ts);
+            else{
                 statement.executeUpdate("TRUNCATE TABLE " + ts);
+            }
         } else {
             //note: if one at a time, need to make sure to first disable FK checks
             for(String t : tablesToClear){
                 if (doDropTable)
-                    statement.execute("DROP TABLE IF EXISTS " +t);
-                else
-                    statement.executeUpdate("TRUNCATE TABLE " + t);
+                    dropTables(statement, t);
+                else{
+                    /*
+                        for MS_SQL_SERVER, we cannot use truncate tables if there exist fk
+                        see
+                            https://docs.microsoft.com/en-us/sql/t-sql/statements/truncate-table-transact-sql?view=sql-server-ver15#restrictions
+                            https://stackoverflow.com/questions/155246/how-do-you-truncate-all-tables-in-a-database-using-tsql#156813
+                        then it will cause a problem to reset identify
+                    */
+                    if (type  == DatabaseType.MS_SQL_SERVER)
+                        deleteTables(statement, t, tablesHaveIdentifies);
+                    else
+                        truncateTables(statement, t);
+                }
             }
         }
+    }
+
+
+    private static void dropTables(Statement statement, String table) throws SQLException {
+        statement.executeUpdate("DROP TABLE IF EXISTS " +table);
+    }
+
+    private static void deleteTables(Statement statement, String table, Set<String> tableHasIdentify) throws SQLException {
+        statement.executeUpdate("DELETE FROM "+table);
+//        NOTE TAHT ideally we should reseed identify here, but there would case an issue, i.e., does not contain an identity column
+        if (tableHasIdentify.contains(table))
+            statement.executeUpdate("DBCC CHECKIDENT ('"+table+"', RESEED, 0)");
+    }
+
+    private static void truncateTables(Statement statement, String table) throws SQLException {
+        statement.executeUpdate("TRUNCATE TABLE " + table);
     }
 
     private static void resetSequences(Statement s, DatabaseType type, String schemaName) throws SQLException {
         ResultSet rs;// Idem for sequences
         Set<String> sequences = new HashSet<>();
+
         rs = s.executeQuery(getAllSequenceCommand(type, schemaName));
         while (rs.next()) {
             sequences.add(rs.getString(1));
@@ -177,6 +222,12 @@ public class DbCleaner {
         switch (type)
         {
             case POSTGRES: break;
+            case MS_SQL_SERVER:
+                //https://stackoverflow.com/questions/159038/how-can-foreign-key-constraints-be-temporarily-disabled-using-t-sql
+                //https://stackoverflow.com/questions/155246/how-do-you-truncate-all-tables-in-a-database-using-tsql#156813
+                //https://docs.microsoft.com/en-us/sql/relational-databases/tables/disable-foreign-key-constraints-with-insert-and-update-statements?view=sql-server-ver15
+                s.execute("EXEC sp_MSForEachTable \"ALTER TABLE ? NOCHECK CONSTRAINT all\"");
+                break;
             case H2:
                 s.execute("SET REFERENTIAL_INTEGRITY FALSE");
                 break;
@@ -193,6 +244,9 @@ public class DbCleaner {
         switch (type)
         {
             case POSTGRES: break;
+            case MS_SQL_SERVER:
+                s.execute("exec sp_MSForEachTable \"ALTER TABLE ? WITH CHECK CHECK CONSTRAINT all\"");
+                break;
             case H2:
                 /*
                     For H2, we have to delete tables one at a time... but, to avoid issues
@@ -211,6 +265,7 @@ public class DbCleaner {
 
     private static int getDefaultReties(DatabaseType type){
         switch (type){
+            case MS_SQL_SERVER:
             case POSTGRES: return 0;
             case H2:
             case MARIADB:
@@ -222,9 +277,11 @@ public class DbCleaner {
     private static String getDefaultSchema(DatabaseType type){
         switch (type){
             case H2: return "PUBLIC";
+            case MS_SQL_SERVER: return "";
             case MARIADB:
-            case MYSQL: throw new IllegalArgumentException("there is no default schema for MySQL");
+            case MYSQL: throw new IllegalArgumentException("there is no default schema for "+type+", and you must specify a db name here");
             case POSTGRES: return "public";
+//            case MS_SQL_SERVER_2000: return "U";
         }
         throw new DbUnsupportedException(type);
     }
@@ -233,27 +290,45 @@ public class DbCleaner {
         return type == DatabaseType.POSTGRES;
     }
 
-
-    private static String getAllTableCommand(String schema) {
-        return "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES where TABLE_SCHEMA='" + schema + "' AND (TABLE_TYPE='TABLE' OR TABLE_TYPE='BASE TABLE')";
+    private static String getAllTableHasIdentify(DatabaseType type, String schema){
+        if(type != DatabaseType.MS_SQL_SERVER)
+            throw new IllegalArgumentException("getAllTableHasIdentify only supports for MS_SQL_SERVER, not for "+type);
+        return getAllTableCommand(type, schema) + " AND OBJECTPROPERTY(OBJECT_ID(TABLE_NAME), 'TableHasIdentity') = 1";
     }
 
-    private static String getAllSequenceCommand(DatabaseType type, String schemaName)
-    {
+    private static String getAllTableCommand(DatabaseType type, String schema) {
+        switch (type){
+            case MS_SQL_SERVER:
+                // https://stackoverflow.com/questions/175415/how-do-i-get-list-of-all-tables-in-a-database-using-tsql
+                if (schema.isEmpty())
+                    return "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES where TABLE_TYPE='BASE TABLE'";
+                else
+                    return "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES where TABLE_TYPE='BASE TABLE' AND TABLE_CATALOG='"+schema+"'";//schema here should be dbname
+            case MYSQL:
+            case MARIADB:
+            case H2:
+            case POSTGRES:
+                return "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES where TABLE_SCHEMA='" + schema + "' AND (TABLE_TYPE='TABLE' OR TABLE_TYPE='BASE TABLE')";
+        }
+        throw new DbUnsupportedException(type);
+//            case MS_SQL_SERVER_2000:
+//                return "SELECT sobjects.name FROM sysobjects sobjects WHERE sobjects.xtype = '"+schema+"'";
+    }
+
+    private static String getAllSequenceCommand(DatabaseType type, String schemaName) {
         switch (type){
             case MYSQL:
-            case MARIADB: return getAllTableCommand(schemaName);
+            case MARIADB: return getAllTableCommand(type, schemaName);
             case H2:
-            case POSTGRES: return getAllSequenceCommand(getDefaultSchema(type));
+            case POSTGRES: return "SELECT SEQUENCE_NAME FROM INFORMATION_SCHEMA.SEQUENCES WHERE SEQUENCE_SCHEMA='" + schemaName + "'";
+            //https://docs.microsoft.com/en-us/sql/relational-databases/system-information-schema-views/sequences-transact-sql?view=sql-server-ver15
+            case MS_SQL_SERVER: return "SELECT SEQUENCE_NAME FROM INFORMATION_SCHEMA.SEQUENCES";
         }
         throw new DbUnsupportedException(type);
     }
 
 
-    private static String getAllSequenceCommand(String schema)
-    {
-        return "SELECT SEQUENCE_NAME FROM INFORMATION_SCHEMA.SEQUENCES WHERE SEQUENCE_SCHEMA='" + schema + "'";
-    }
+
 
     private static String resetSequenceCommand(String sequence, DatabaseType type) {
         switch (type){
